Revised check for locally cached base images

+++ build-image.sh
@@ -8,6 +8,7 @@ echo "Cloning branch ${GATEWAY_BRANCH} of ${GATEWAY_REPO}"
 set -x
 
 # Do a shallow clone of the gateway repository.
+rm -rf $(basename ${GATEWAY_REPO})
 git clone --depth 1 --branch ${GATEWAY_BRANCH} ${GATEWAY_REPO}
 
 # Grab the latest open-zwave and unpack it
@@ -28,19 +29,24 @@ ${RPXC} bash -c ./gateway/image/build-gateway.sh
 BASE_IMAGE_ZIP="${BASE_IMAGE_NAME}.zip"
 
 mkdir -p base
-rm -f "base/${BASE_IMAGE_ZIP}.sha256sum"
-wget --directory-prefix=base "${BASE_IMAGE_DIR}/${BASE_IMAGE_ZIP}.sha256sum"
 if [ ${USE_CACHED_BASE} == 1 ]; then
-  if [ -f "base/${BASE_IMAGE_ZIP}" ]; then
-    echo "Checking cached base image"
-    (cd base; sha256sum --check "${BASE_IMAGE_ZIP}.sha256sum" || rm -f "${BASE_IMAGE_ZIP}")
-    echo "Checksum verified - using cached base"
-  else
-    echo "No cached copy - downloading base"
+  if [ ! -f "base/${BASE_IMAGE_ZIP}.sha256sum" ]; then
+    wget --directory-prefix=base "${BASE_IMAGE_DIR}/${BASE_IMAGE_ZIP}.sha256sum"
+    if [ -f "base/${BASE_IMAGE_ZIP}" ]; then
+      echo "Checking cached base image"
+      (cd base; sha256sum --check "${BASE_IMAGE_ZIP}.sha256sum" || rm -f "${BASE_IMAGE_ZIP}")
+      echo "Checksum verified - using cached base"
+    else
+      echo "No cached copy - downloading base"
+    fi
   fi
+else
+  rm -f "base/${BASE_IMAGE_ZIP}"
+  rm -f "base/${BASE_IMAGE_ZIP}.sha256sum"
 fi
 if [ ! -f "base/${BASE_IMAGE_ZIP}" ]; then
   wget --directory-prefix=base "${BASE_IMAGE_DIR}/${BASE_IMAGE_ZIP}"
+  wget --directory-prefix=base "${BASE_IMAGE_DIR}/${BASE_IMAGE_ZIP}.sha256sum"
 fi
 (cd base; sha256sum --check "${BASE_IMAGE_ZIP}.sha256sum")
